+++ CHANGELOG.md
@@ -14,7 +14,13 @@
 
 -->
 
-## 1.4.6
+### Master
+
+### 1.4.7
+
+* Fixes crash when following an artist from the ‘related artist’ Home/ForYou rail - alloy
+
+### 1.4.6
 
 * Scroll to last sent message upon send - sarah
 * Upgrade to latest Relay, including our custome Node ID and language plugin patches - alloy